+++ libs/JSystem/src/JKernel/JKRDvdRipper.cpp
@@ -55,7 +55,7 @@ void* JKRDvdRipper::loadToMainRAM(JKRDvdFile* dvdFile, u8* dst,
 	s32 fileSizeAligned = ALIGN_NEXT(dvdFile->getFileSize(), 32);
 	if (expandSwitch == EXPAND_SWITCH_DECOMPRESS) {
 		u8 buffer[0x40];
-		u8* bufPtr = (u8*)ALIGN_NEXT((u32)buffer, 32);
+		u8* bufPtr = (u8*)ALIGN_NEXT((uintptr_t)buffer, 32);
 		while (true) {
 			int readBytes
 			    = DVDReadPrio(dvdFile->getFileInfo(), bufPtr, 0x20, 0, 2);
@@ -125,7 +125,7 @@ void* JKRDvdRipper::loadToMainRAM(JKRDvdFile* dvdFile, u8* dst,
 
 		if (offset != 0) {
 			u8 buffer[0x40];
-			u8* bufPtr = (u8*)ALIGN_NEXT((u32)buffer, 32);
+			u8* bufPtr = (u8*)ALIGN_NEXT((uintptr_t)buffer, 32);
 			while (true) {
 				int readBytes = DVDReadPrio(dvdFile->getFileInfo(), bufPtr, 32,
 				                            (s32)offset, 2);
