+++ src/Map/MapMirror.cpp
@@ -306,7 +306,7 @@ void TMirrorModelManager::loadAfter()
 
 		target = source;
 		target.imageDataOffset
-		    = (u32)&source + source.imageDataOffset - (u32)&target;
+		    = (uintptr_t)&source + source.imageDataOffset - (uintptr_t)&target;
 	}
 }
